fix(nbvm): sync KNOWN_GLOBALS with the installed global set (+ wasm32 support)
Groundwork for a browser demo turned up three real defects.

**`typeof Temporal` was `"undefined"`.** `KNOWN_GLOBALS` is a hand-kept
mirror of the globals the interpreter installs, and it had gone stale by
twelve names — `Temporal`, `Atomics`, `SharedArrayBuffer`,
`Float16Array`, `FinalizationRegistry`, `DisposableStack`,
`AsyncDisposableStack`, `ShadowRealm`, `SuppressedError`, `atob`, `btoa`,
`eval`. A missing entry is wrong twice over: the value path emits a
spurious `ReferenceError`, and `typeof` is *answered at compile time* as
`"undefined"` for something that plainly exists. Add the missing names,
and add `globals_match_installed_set`, which diffs the list against
`Object.getOwnPropertyNames(globalThis)` in both directions so it cannot
drift again.

**The crate did not build for `wasm32-unknown-unknown`.** A `usize`
literal of 2^53-1 does not fit in 32 bits; hold it as `u64` (the limit is
unreachable on a 32-bit target anyway).

**`SystemTime::now` panics on bare wasm.** `std` exists there but has no
time source, so seeding the RNG or reading the clock aborted the module.
The engine already treats "no clock" as a supported case for `no_std`;
extend that to `wasm32-unknown-unknown` specifically (wasip1 and the
JS-glue targets do have a clock, so the exclusion stays narrow).

Also adds three C ABI entry points the browser needs: `kt_alloc`/
`kt_free`, because a host that cannot allocate inside the module's
address space must ask it for a region first, and `kt_eval_capture`,
which returns the script's console output alongside its completion value
(NUL-separated) — `kt_eval` reports only the latter, which loses
everything a script prints.

+++ src/nbexec/method_dispatch.rs
@@ -5481,7 +5481,9 @@ impl<'a> Interp<'a> {
         mapper: Option<NanBox>,
         this_arg: NanBox,
     ) -> Result<usize, ExecError> {
-        const MAX_SAFE: usize = 9_007_199_254_740_991; // 2^53 - 1
+        // 2^53 - 1. Held as `u64` because `usize` is 32 bits on wasm32, where
+        // the literal does not fit (and the limit is unreachable anyway).
+        const MAX_SAFE: u64 = 9_007_199_254_740_991;
         let mut target_index = start;
         for source_index in 0..source_len {
             let key = alloc::format!("{source_index}");
@@ -5515,7 +5517,7 @@ impl<'a> Interp<'a> {
                     NanBox::undefined(),
                 )?;
             } else {
-                if target_index >= MAX_SAFE {
+                if target_index as u64 >= MAX_SAFE {
                     return Err(self.type_error("flatten result exceeds maximum array length"));
                 }
                 self.create_data_property_or_throw(target, target_index, element)?;
